requirements.txt
torch
torchvision
matplotlib
numpy
streamlit
streamlit-drawable-canvas
psycopg2-binary
Pillow